refactor(api): extract vault secret kernel and unify byte encoding
The AES-GCM envelope-encryption kernel lived in modules/mcp while
channels, environments, and vendor-credentials reached into MCP
internals for crypto. Move it to modules/vault/application/
vault-secret-store.ts, matching the Credential / Secret Vault Service
boundary in docs/architecture.md. requireStateSecret stays private to
mcp-oauth-state.service, its only caller.

Consolidate the byte-encoding helpers (toArrayBuffer defined 4x,
base64url defined 4x across mcp-secret-store, response-body,
runtime-boot-token, cli-oauth-device, personal-access-token) into
shared/bytes.ts and delete shared/response-body.ts.
