+++ apps/api/src/modules/channels/wechat/wechat-context-token-secret-store.ts
@@ -5,7 +5,7 @@ import { and, eq } from "drizzle-orm";
 import type { ApiBindings } from "../../../platform/cloudflare/worker-types";
 import { getAppDatabase } from "../../../platform/db/drizzle";
 import { validationError } from "../../../platform/errors";
-import { readSecretOutcome, storeSecret } from "../../mcp/application/mcp-secret-store";
+import { readSecretOutcome, storeSecret } from "../../vault/application/vault-secret-store";
 
 export type WeChatContextTokenSecretReadPurpose = "wechat_reply";
